move scripts/utils.js to root
